1.5.2 - changed Node into private class
