+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -38,4 +38,34 @@ public class IntLinkedList {
         return ret;
     }
 
+    private static class Node {
+
+        private int value;
+        private Node prev;
+        private Node next;
+
+        public Node(int i) {
+            value = i;
+        }
+
+        public int getValue() {
+            return value;
+        }
+
+        public Node getPrev() {
+            return prev;
+        }
+
+        public void setPrev(Node prev) {
+            this.prev = prev;
+        }
+
+        public Node getNext() {
+            return next;
+        }
+
+        public void setNext(Node next) {
+            this.next = next;
+        }
+    }
 }
